pyproject.toml
[tool.poetry]
name        = "mgraph_ai_service_html_graph"
version     = "v1.0.16"
description = "MGraph-AI__Service__Html__Graph"
authors     = ["Dinis Cruz <dinis.cruz@owasp.org>"]
license     = "Apache 2.0"
readme      = "README.md"
homepage    = "https://github.com/the-cyber-boardroom/MGraph-AI__Service__Html__Graph"
repository  = "https://github.com/the-cyber-boardroom/MGraph-AI__Service__Html__Graph"


packages    = [
    { include = "mgraph_ai_service_html_graph"            },
    { include = "mgraph_ai_service_html_graph__render_ui" }
]

[tool.poetry.dependencies]
python                     = "^3.12"
osbot-fast-api-serverless  = "*"
mgraph_db                  = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"